adds algorithm for cal count for routes

+++ app/src/main/java/com/mikeschen/www/fitnessapp/MainActivity.java
@@ -188,7 +188,7 @@ public class MainActivity extends AppCompatActivity implements
     }
 
     @Override
-    public void showCalorieRoute(double calorie) {
+    public void showCalorieRoute(int calorie) {
     }
 
     @Override
