+++ app/src/main/java/com/mikeschen/www/fitnessapp/MapInterface.java
@@ -10,7 +10,7 @@ public interface MapInterface {
         void showMap();
         void showDistance(String distance);
         void showDuration(String duration);
-        void showCalorieRoute(double calorie);
+        void showCalorieRoute(int calorie);
     }
 
     interface Presenter {
